src/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
antlr4-python3-runtime==4.8
archspec @ file:///croot/archspec_1709217642129/work
asttokens==3.0.0
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitarray==3.2.0
bleach==6.2.0
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1717618050233/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
conda @ file:///croot/conda_1715635703388/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
contourpy==1.3.0
cryptography @ file:///croot/cryptography_1714660666131/work
cuda-python==12.5.0
cycler==0.12.1
Cython==3.0.12
decorator==5.1.1
defusedxml==0.7.1
distro @ file:///croot/distro_1714488253808/work
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.0
filelock==3.15.4
fonttools==4.54.1
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.4.1
fsspec==2024.6.1
gevent==24.2.1
geventhttpclient==2.0.2
greenlet==3.0.3
grpcio==1.65.1
huggingface-hub==0.29.3
hydra-core==1.0.7
idna @ file:///croot/idna_1714398848350/work
imageio==2.34.2
importlib_metadata==8.2.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lazy_loader==0.4
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
lxml==5.3.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1718132535249/work
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvidia-pytriton==0.5.10
omegaconf==2.0.6
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
portalocker==3.1.1
prompt_toolkit==3.0.48
protobuf==4.25.4
ptyprocess==0.7.0
pure_eval==0.2.3
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycuda==2024.1.2
Pygments==2.18.0
nvidia-ml-py==12.560.30
pyparsing==3.2.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-rapidjson==1.18
pytools==2024.1.13
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
referencing==0.35.1
regex==2024.11.6
requests @ file:///croot/requests_1716902831423/work
rich==13.7.1
rpds-py==0.21.0
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
sacrebleu==2.5.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.14.0
setuptools==69.5.1
sh==2.0.7
shellingham==1.5.4
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4.2
tqdm @ file:///croot/tqdm_1716395931952/work
traitlets==5.14.3
transformers==4.50.3
triton==3.2.0
tritonclient==2.48.0
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1718912636303/work
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
zstandard @ file:///croot/zstandard_1714677652653/work
